BookRepositoryTest done

+++ src/main/java/s24/backend/bookstore/domain/Book.java
@@ -39,7 +39,7 @@ public class Book {
     public Book() {
     }
 
-    public Book(String title, String author, int publicationYear, String isbn, int price, Category category) {
+    public Book(String title, String author, int publicationYear, String isbn, double price, Category category) {
         this.title = title;
         this.author = author;
         this.publicationYear = publicationYear;
@@ -48,13 +48,14 @@ public class Book {
         this.category = category;
     }
 
-    public Book(String title, String author, int publicationYear, String isbn, int price) {
+    public Book(String title, String author, int publicationYear, String isbn, double price) {
         this.title = title;
         this.author = author;
         this.publicationYear = publicationYear;
         this.isbn = isbn;
         this.price = price;
     }
+
     public String getTitle() {
         return title;
     }
